requirements.txt
amqp==5.2.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
packaging==24.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
python-arango==7.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.20.0
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stix2==3.0.1
stix2-patterns==2.0.0
stix2arango @ https://github.com/muchdogesec/stix2arango/archive/main.zip
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
arango_cti_processor @ https://github.com/muchdogesec/arango_cti_processor/archive/main.zip#39
